// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const FileStatus = {
  uploading: 'uploading',
  ready: 'ready',
  deleting: 'deleting',
  deleted: 'deleted',
  failed: 'failed',
  missing: 'missing'
} as const

export type FileStatus = (typeof FileStatus)[keyof typeof FileStatus]


export const OptimizationStatus = {
  pending: 'pending',
  processing: 'processing',
  ready: 'ready',
  failed: 'failed'
} as const

export type OptimizationStatus = (typeof OptimizationStatus)[keyof typeof OptimizationStatus]